\name{geneAnnotation}
\docType{data}
\alias{geneAnnotation}
\title{geneAnnotation}
\description{
    The object contains a data.frame with Affymetrix transcript cluster id and
    gene symbol for the 22,011 genes defined by the Affymetrix core gene set.
}
\usage{
    data(geneAnnotation)
}
\value{
    A data.frame with Affymetrix transcript cluster id and gene symbol for
    22,011 genes.
}
\format{data.frame with two columns; Affymetrix transcript cluster id and 
    gene symbol.
}
