\name{geneSets}
\docType{data}
\alias{geneSets}
\title{geneSets}
\description{
    A data.frame with 1,454 lists of genes (specified by Affymetrix
    transcript cluster id's and gene symbols) to be included in the 
    correlation analysis. The data.frame contains a complete
    collection of all Gene Ontology gene sets included in the
    Molecular Signatures Database v3.1.
}
\usage{
    data(geneSets)
}
\value{
    A data.frame with 1,454 Gene Ontology gene lists specified by
    Affymetrix transcript cluster id's and gene symbols.
}
\format{data.frame that contains 1,454 lists of genes, specified by
    Affymetrix transcript cluster id's and gene symbols.
}
