\name{sampleSetFirmaScores}
\docType{data}
\alias{sampleSetFirmaScores}
\title{sampleSetFirmaScores}
\description{
    A data.frame containing preprocessed log2 expression values from 16 samples
    in 10,000 randomly selected Affymetrix probesets.
}
\usage{
    data(sampleSetFirmaScores)
}
\value{
    A data.frame with preprocessed log2 expression values in 10,000 randomly
    selected Affymetrix probesets for a test data set of 16 samples.
}
\format{data.frame that contains preprocessed log2 expression values from 16
    samples in 10,000 randomly selected Affymetrix probesets.
}
