\name{sampleSetGeneSummaries}
\docType{data}
\alias{sampleSetGeneSummaries}
\title{sampleSetGeneSummaries}
\description{
    A data.frame containing gene expression summary values from 16 samples in 
    22,011 Affymetrix transcript clusters. 
}
\usage{
    data(sampleSetGeneSummaries)
}
\value{
    A data.frame with expression values at the gene level for a test data set
    of 16 samples.
}
\format{A data.frame containing gene expression summary values from 16
    samples in 22,011 Affymetrix transcript clusters.
}
