\name{splicingFactors}
\docType{data}
\alias{splicingFactors}
\title{A list of 280 splicing factor genes}
\description{
    The gene set is a comprehensive collection of 280 genes involved
    in pre-mRNA splicing events (Sveen et al., Genome Medicine, 2011, 3:32).
}
\usage{
    data(splicingFactors)
}
\value{
    A data.frame with 280 genes involved in pre-mRNA splicing events.
}
\format{data.frame with two columns; Affymetrix transcript cluster id
    and gene symbol.
}
